requirements.txt
fastapi
typing-extensions
fastapi[standard]
PyJWT
passlib[bcrypt]
python-dotenv
sqlalchemy